% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_completion_rate}
\alias{compute_completion_rate}
\title{Compute completion rate}
\usage{
compute_completion_rate(data, id_vars = NULL, digits = 2)
}
\arguments{
\item{data}{A tibble.}

\item{id_vars}{A character vector of variable names indicating the groups
over which counting of the \code{NA} should be performed. Default
to \code{NULL} indicating that no groups should be used.}

\item{digits}{The \code{digits} argument passed to \code{\link[=signif]{signif()}}.}
}
\value{
A tibble.
}
\description{
compute_completion_rate() computes
}
\examples{
\dontrun{
data_list <- load_dx("tb")
dm <- build_dm(data_list, year = 2019)
tbl <- build_tbl(dm)
compute_completion_rate(tbl, id_vars = c("year", "is_hbc"))

cars_df <- tibble::as_tibble(mtcars, rownames = "car_name")
compute_completion_rate(cars_df, id_vars = NULL)
}
}
