% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute_correlation}
\alias{compute_correlation}
\title{Compute correlations with a target variable}
\usage{
compute_correlation(data, target_var, by = NULL, ...)
}
\arguments{
\item{data}{A tibble.}

\item{target_var}{The variable against which the correlations needs to be
computed. For instance, \code{dx_gap}.}

\item{by}{A character vector.}

\item{...}{Optional arguments passed to \code{\link[=correlate]{correlate()}}}
}
\value{
A tibble.
}
\description{
\code{compute_correlation()} allows to compute correlation of several numerical
predictors with respect to a target variable. Optionally, the computation
can be done by groups.
}
\examples{
\dontrun{
tbl <-
  build_tbl(
    "tb",
    year = NULL,
    estimated = "who_estimates.e_inc_num",
    notified = "who_notifications.c_newinc",
    vars = dxgap_const$tb_vars
  ) |>
  dplyr::mutate(is_hbc = forcats::as_factor(is_hbc))

tbl |>
  dplyr::filter(year == 2019) |>
  compute_correlation(dx_gap, by = "year")

tbl |>
  compute_correlation(dx_gap, by = c("year", "is_hbc"))

car_tbl <- tibble::as_tibble(mtcars, rownames = "car_name")
my_cars <- dplyr::select(car_tbl, -dplyr::all_of(c("vs", "am", "carb")))
compute_correlation(my_cars, mpg, by = c("gear", "cyl"))
compute_correlation(my_cars, mpg, by = NULL) # car_name is dropped by default
}
}
