% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dxgap_master_list}
\alias{dxgap_master_list}
\title{Data dictionary}
\format{
\subsection{\code{dxgap_master_list}}{

A data frame:
\describe{
\item{data_source}{The name of data provider}
\item{variable_name}{The variable name}
\item{url_endpoint}{The endpoint if the data set can be fetched programmatically}
\item{definition}{Description of the \code{variable_name}}
\item{url}{Url of the source, if any}
\item{notes}{Additional notes, if any}
}
}
}
\usage{
dxgap_master_list
}
\description{
It shows all supported data of the project. It should be kept up-to-date.
}
\keyword{datasets}
