% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\name{extract_mod_coeff}
\alias{extract_mod_coeff}
\title{Extract coefficients from final fit}
\usage{
extract_mod_coeff(mod_out_list)
}
\arguments{
\item{mod_out_list}{A list object as returned by \code{\link[=run_mod]{run_mod()}}.}
}
\value{
A tibble.
}
\description{
An helper functions to extract coefficient estimates from the best model
returned by \code{\link[=run_mod]{run_mod()}}.
}
