% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{extract_vars}
\alias{extract_vars}
\title{Extract supported variables from \code{dxgap_diseases} meta table}
\usage{
extract_vars(disease, .dxgap_diseases = dxgap_diseases)
}
\arguments{
\item{disease}{The disease identifier.}

\item{.dxgap_diseases}{The \code{dxgap_diseases} table.}
}
\value{
A character vector of supported variables.
}
\description{
Extract supported variables from \code{dxgap_diseases} meta table
}
