% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.dxgap-package.R
\docType{package}
\name{find.dxgap-package}
\alias{find.dxgap}
\alias{find.dxgap-package}
\title{find.dxgap: Compute dxgap metric for different diseases}
\description{
Compute dxgap for different diseases.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://finddx.github.io/find.dxgap/}
}

}
\author{
\strong{Maintainer}: Andrea Ranzato \email{andrea@cynkra.com}

Authors:
\itemize{
  \item Mike John Page \email{mike@cynkra.com}
  \item Anna Mantsoki \email{anna.mantsoki@finddx.org}
}

Other contributors:
\itemize{
  \item FIND [copyright holder, funder]
}

}
\keyword{internal}
