% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\name{get_mod_preproc}
\alias{get_mod_preproc}
\title{Get pre-processing recipes}
\usage{
get_mod_preproc(.tbl, .neighbors, .threshold, .impute_vars)
}
\arguments{
\item{.tbl}{Input data frame containing the data to model.}

\item{.neighbors}{The number of neighbors.}

\item{.threshold}{A value for the threshold of missing values in column. The
step will remove the columns where the proportion of missing values
exceeds the threshold.}
}
\value{
A list of items of class "recipe".
}
\description{
This function returns a list of pre-processing steps applied with
\code{\link[=run_mod]{run_mod()}}.
}
\examples{
\dontrun{
preproc_list <- get_mod_preproc(
  .tbl = tbl,
  .neighbors = 5,
  .threshold = 0.25,
  .impute_vars = c("gdp", "e_inc_num", "pop_total")
 )
}
}
