% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_bulk}
\alias{import_bulk}
\title{Import tables in bulk}
\usage{
import_bulk(data_lst, data_name, data_dir = Sys.getenv("DXGAP_DATADIR"))
}
\description{
Import tables in bulk
}
\examples{
\dontrun{
lst_df <- import_tbl(lst_df, "who_2023-07-28_hbc.csv")
}
}
