% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_tbl}
\alias{import_tbl}
\title{Import table}
\usage{
import_tbl(
  .data_lst,
  .file_name,
  .all = TRUE,
  .data_dir = Sys.getenv("DXGAP_DATADIR")
)
}
\description{
Import table
}
\examples{
\dontrun{
lst_df <- import_tbl(lst_df, "who_2023-07-28_hbc.csv")
}
}
