% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_helpers.R
\name{extract_mod_coeff}
\alias{extract_mod_coeff}
\alias{extract_mod_fit}
\alias{extract_mod_rank}
\title{Extract coefficients from final fit}
\usage{
extract_mod_coeff(mod_out_list)

extract_mod_fit(mod_out_list)

extract_mod_rank(mod_out_list)
}
\arguments{
\item{mod_out_list}{A list object as returned by \code{\link[=run_mod]{run_mod()}}.}
}
\value{
\code{extract_mod_coeff()} returns a tibble.

\code{extract_mod_fit()} returns an object of class \verb{_lm}.

\code{extract_mod_fit()} returns an object of class \verb{_lm}.
}
\description{
\code{extract_mod_coeff()} is an helper functions to extract coefficient estimates
from the best model returned by \code{\link[=run_mod]{run_mod()}}.

\code{extract_mod_fit()} is an helper function to extract the model object from
the best model returned by \code{\link[=run_mod]{run_mod()}}.

\code{extract_mod_fit()} is an helper function to extract the model object from
the best model returned by \code{\link[=run_mod]{run_mod()}}.
}
\examples{
\dontrun{
mod_objects <- run_mod(
  tbl,
  preproc = preproc_list,
  mod = mod_list,
  folds = tb_mod_const$folds,
  metrics = tb_mod_const$metrics,
  rank_metric = tb_mod_const$rank_metric,
)
extract_mod_coeff(mod_objects)
}
\dontrun{
lm_obj <- extract_mod_fit(mod_objects)
plot_check <- performance::check_model(
  lm_obj,
  check = c("linearity", "normality", "qq", "outliers"),
  theme = "ggplot2::theme_minimal"
)
}
\dontrun{
lm_obj <- extract_mod_fit(mod_objects)
plot_check <- performance::check_model(
  lm_obj,
  check = c("linearity", "normality", "qq", "outliers"),
  theme = "ggplot2::theme_minimal"
)
}
}
