% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_tb_ts}
\alias{plot_tb_ts}
\title{Plot time series}
\usage{
plot_tb_ts(
  metric = "tb_incidence_per_100k",
  years = seq(2015, 2020),
  countries = c("Angola", "Brazil", "Zambia")
)
}
\arguments{
\item{metric}{A character of length one of the indicator to display.}

\item{years}{A numeric vector for the years of the time series.}

\item{countries}{A character vector with country names to display.}
}
\value{
A plot.
}
\description{
\code{plot_tb_ts()} renders a time series for given indicator, years and countries
chosen by the user. Input data for the time series are read and tidied up by
by \code{read_tb_dashboard()} and \code{tidy_tb_dashboard()}.
}
\examples{
\dontrun{
plot_tb_ts()
}
}
