% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_tb_dashboard}
\alias{read_tb_dashboard}
\title{Read the TB dashboard}
\usage{
read_tb_dashboard()
}
\value{
A tibble
}
\description{
\code{read_tb_dashboard()} reads the raw Excel TB dashboard into R, keeping all
the information in an untidied format.
}
\examples{
\dontrun{
read_tb_dashboard()
}
}
\seealso{
\code{\link[=tidy_tb_dashboard]{tidy_tb_dashboard()}} to tidy this dashboard.
}
