% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{render_bulk}
\alias{render_bulk}
\title{Render template across multiple years}
\usage{
render_bulk(template_name, year, vars)
}
\description{
Render template across multiple years
}
\examples{
\dontrun{
render_bulk("eda.Rmd", year = 2018:2021, vars = dxgap_constants$tb_vars)
}
}
