% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{render_report}
\alias{render_report}
\title{Render template}
\usage{
render_report(
  .template_name,
  .year = NULL,
  .vars = NULL,
  interactive = TRUE,
  template_dir = "inst/template/",
  data_dir = Sys.getenv("DXGAP_DATADIR")
)
}
\description{
Render template
}
\examples{
\dontrun{
render_report("eda.Rmd", .year = 2019, .vars = dxgap_const$tb_vars)
}
}
