% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\name{run_mod}
\alias{run_mod}
\title{Run a suite of statistical models}
\usage{
run_mod(
  tbl,
  mod,
  preproc,
  folds,
  metrics,
  rank_metric,
  cross = TRUE,
  seed = 2023
)
}
\arguments{
\item{tbl}{Input data frame containing the data to model.}

\item{mod}{A list of models to run.}

\item{preproc}{A list of preprocessing steps.}

\item{folds}{An integer. The number of cross-validation folds.}

\item{metrics}{A tibble containing the performance metrics to evaluate.}

\item{rank_metric}{A metric from \code{metrics} to rank results by.}

\item{cross}{A logical: should all combinations of the preprocessors and
models be used to create the workflows? If FALSE, the length of preproc and
models should be equal.}

\item{seed}{A single integer.}
}
\value{
A list containing a workset, ranked model results, and a final model
fit.
}
\description{
\code{run_mod()} runs a suite of statical models, returning a final model fit.
}
\examples{
\dontrun{
tbl <- load_dx() |>
  build_dm(year = 2019) |>
  build_tbl(vars = dxgap_const$tb_vars) |>
  compute_dx_gap() |>
  dplyr::mutate(is_hbc = forcats::as_factor(is_hbc)) |>
  dplyr::select(-any_of(c("year", "country")))

 preproc_list <- get_mod_preproc(
  .tbl = tbl,
  .neighbors = tb_mod_const$neighbors,
  .threshold = tb_mod_const$threshold,
  .impute_vars = tb_mod_const$impute_vars
 )

mod_list <- get_mod_mod(tb_mod_const$mode, tb_mod_const$engine)

run_mod(
  tbl,
  preproc = preproc_list,
  mod = mod_list,
  folds = tb_mod_const$folds,
  metrics = tb_mod_const$metrics,
  rank_metric = tb_mod_const$rank_metric,
)
}
}
