% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_findtb_datadir}
\alias{set_findtb_datadir}
\title{Create folder containing data for the project.}
\usage{
set_findtb_datadir(
  path = fs::path_dir(getwd()),
  dirs = c("who", "global-fund", "world-bank", "meta")
)
}
\arguments{
\item{path}{a path where the folder should be created. By defeult, it is
created at the same level of where the findtb project lives.}

\item{dirs}{a character vector with subfolders' names.}
}
\value{
the path to the folder, invisibly.
}
\description{
This function has side effects. It creates a folder storing data for the
project. In addition it sets the \code{FINDTB_DATADIR} environment variable which
is needed for the \verb{read_*()} family of functions to work properly, in
particular when running tests.
}
\examples{
path_findtb <- set_findtb_datadir("~/Documents/my_projects")
}
