% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashb.R
\name{tidy_tb_dashboard}
\alias{tidy_tb_dashboard}
\title{Tidy the TB dashboard}
\usage{
tidy_tb_dashboard(data, type = c("time_series", "fixed"))
}
\arguments{
\item{data}{Input data set. Must come from \code{\link[=read_tb_dashboard]{read_tb_dashboard()}}.}

\item{type}{Flag to indicate whether time series or fixed data should be
returned. Defaults to \code{time_series}.}
}
\value{
A tibble
}
\description{
\code{tidy_tb_dashboard()} tidies the messy TB dashboard read by
\code{\link[=read_tb_dashboard]{read_tb_dashboard()}}. The messy data combines both time series and fixed
data so a flag is provided to select the type of data that should be
returned, with a shared 'country' primary key.
}
\examples{
\dontrun{
# Returns the time series data by default
tidy_tb_dashboard()

# Return fixed data
# tidy_tb_dashboard(time_series = FALSE)
}
}
\seealso{
\code{\link[=read_tb_dashboard]{read_tb_dashboard()}}
}
