% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb.R
\name{wb}
\alias{wb}
\alias{download_wb}
\title{Data from World Bank}
\usage{
download_wb(file_name, indicator, range_years)
}
\arguments{
\item{indicator}{A string indicating the label of the data set as documented
in the World Bank API. For instance, "SP.POP.TOTL".}

\item{range_years}{The range of the years to be downloaded as a string.}
}
\value{
\code{download_wb()} returns invisibly the file path in which data are
stored.
}
\description{
These help pages document the lower-level API to individually download, read,
and tidy data. For a higher-level API that works across all data sets,
see:
\itemize{
\item \code{\link[=import_tbl]{import_tbl()}} to download all data
}
}
\details{
The data sets currently available from World Bank in this package are:
\itemize{
\item urban population
\item total population
\item population density
\item gdp
}
}
\section{Source}{

The World Bank API is documented at \url{https://datahelpdesk.worldbank.org/knowledgebase/articles/898581}.
}

\examples{
\dontrun{
pop_urban <- download_wb(
  file_name = paste_dataset_name_date("wb", dataset = "pop_urban", file_ext = ".csv"),
  indicator = "SP.URB.TOTL.IN.ZS",
  range_years = "2015:2023"
)
pop_density <-
  download_wb(
    file_name = paste_dataset_name_date("wb", dataset = "pop_density", file_ext = ".csv"),
    indicator = "EN.POP.DNST",
    range_years = "2015:2023"
  )
pop_total <- download_wb(
  file_name = paste_dataset_name_date("wb", dataset = "pop_total", file_ext = ".csv"),
  indicator = "SP.POP.TOTL",
  range_years = "2015:2023"
)
gdp <- download_wb(
  file_name = paste_dataset_name_date("wb", dataset = "gdp", file_ext = ".csv"),
  indicator = "NY.GDP.MKTP.CD",
  range_years = "2015:2023"
)
}
}
