% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who.R
\name{who}
\alias{who}
\alias{download_who}
\title{Data from WHO}
\usage{
download_who(file_name, url_endpoint)
}
\arguments{
\item{file_name}{A path to the file. By default, the file is downloaded into
a temporary location (useful when running unit tests). Instead, a custom
target directory can be set using an absolute path, for instance,
"~/path_to/my_file.ext". In both cases the path specified in \code{data_dir}
does not play a role. Alternatively, if a relative path is provided,
function downloads the data into the the default folder "find.dxgap.data".
Please, make sure the environment variable "DXGAP_DATADIR" is set in your
.Renviron.}

\item{url_endpoint}{A string indicating the name of the data set to be
downloaded. One of the following:
\itemize{
\item budget
\item community
\item estimates
\item expenditure_utilisation
\item labs
\item notifications
}}
}
\value{
\code{download_who()} returns invisibly the file path in which data are
stored.
}
\description{
These help pages document the lower-level API to individually download, read,
and tidy data. For a higher-level API that works across all data sets,
see:
\itemize{
\item \code{\link[=import_tbl]{import_tbl()}} to download all data
}
}
\details{
The data sets currently available from WHO in this package are:
\itemize{
\item notifications
\item estimates
\item budget
\item community
\item expenditures
\item laboratories
\item hbc
}
}
\examples{
\dontrun{
notification <- download_who(
  file_name = paste_dataset_name_date("who", dataset = "notifications", file_ext = ".csv"),
  url_endpoint = "notifications"
)
estimates <- download_who(
  file_name = paste_dataset_name_date("who", dataset = "estimates", file_ext = ".csv"),
  url_endpoint = "estimates"
)
budget <- download_who(
  file_name = paste_dataset_name_date("who", dataset = "budget", file_ext = ".csv"),
  url_endpoint = "budget"
)
community_engagement <- download_who(
  file_name = paste_dataset_name_date("who", dataset = "community", file_ext = ".csv"),
  url_endpoint = "community"
)
expenditure_and_utilisation <- download_who(
  file_name = paste_dataset_name_date("who", dataset = "expenditures", file_ext = ".csv"),
  url_endpoint = "expenditure_utilisation"
)
laboratories <- download_who(
  file_name = paste_dataset_name_date("who", dataset = "laboratories", file_ext = ".csv"),
  url_endpoint = "labs"
)
}
}
