% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{write_data_dir}
\alias{write_data_dir}
\title{Create a folder containing data for the project.}
\usage{
write_data_dir(path = fs::path_dir(getwd()), dirs = character(0))
}
\arguments{
\item{path}{A path where the folder should be created. By default, it is
created at the same level as the find.dxgap root directory.}

\item{dirs}{Defaults to \code{character(0)}. An optional character vector to build
additional sub-folders. Not yet in use.}
}
\value{
the path to the folder, invisibly.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Data for the project are now stored in a repository called
\href{https://github.com/finddx/find.dxgap.data}{\code{find.dxgap.data}}.
}
\details{
This function has side effects. It creates a folder storing data for the
project. In addition it sets the \code{DXGAP_DATADIR} environment variable which
is needed for the \verb{read_*()} family of functions to work, in particular when
running tests.
}
\examples{
\dontrun{
write_data_dir() # Create folder at same level as project
write_data_dir("~/Documents/my_projects") # Specify specific folder
}
}
\keyword{internal}
