% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{compute_dx_gap}
\alias{compute_dx_gap}
\title{Compute the diagnostic gap}
\usage{
compute_dx_gap(data)
}
\arguments{
\item{data}{Input data. It assumes the variables \code{e_inc_num} and \code{c_newinc}
are present in the data, describing the number of estimated and notified
cases respectively.}
}
\value{
A tibble, the same dimensions as the input data, but with one
additional column called \code{who_dx_gap} containing the computed diagnostic
gap values
}
\description{
\code{compute_dx_gap()} calculates a diagnostic gap defined as the proportion of
the population with the condition who are undiagnosed.
}
\section{Calculation}{


For each country and year in the data, the diagnostic gap can be computed
as follows:

diagnostic gap = (estimated cases - notified cases) / estimated cases

In other words, we can define the diagnostic gap as the gap between the
estimated number of people who fall ill each year and the number of people
who are officially reported as having been diagnosed. This means that the
diagnostic gap represents people who have the illness but are not
officially diagnosed, due to limitations or challenges in the healthcare
system.
}

\examples{
\dontrun{
# Calculate diagnostic gap for 2019 TB data:
load_dx() |>
 build_dm(year = 2019) |>
 build_tbl_impl(vars = dxgap_const$tb_vars) |>
 compute_dx_gap()
}
}
