% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gf.R
\name{gf}
\alias{gf}
\title{Data from Global Fund}
\description{
These help pages document the lower-level API to individually download, read,
and tidy Global Fund data. For a higher-level API that works across all data
sets, see:
\itemize{
\item \code{\link[=import_tbl]{import_tbl()}} to download all data
}
}
\details{
The data sets currently supported from Global Fund in this package are:
\itemize{
\item procurement
}
}
\section{Source}{

\url{https://insights.theglobalfund.org/t/Public/views/PriceQualityReportingTransactionSummary/TransactionSummary?iframeSizedToWindow=true&\%3Aembed=y&\%3AshowAppBanner=false&\%3Adisplay_count=no&\%3AshowVizHome=no?}
}

