% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who_hbc.R
\name{hbc}
\alias{hbc}
\alias{download_hbc}
\title{TB High-Burden Countries data from WHO}
\usage{
download_hbc(file_name)
}
\arguments{
\item{file_name}{A path to the file. By default, the file is downloaded into
a temporary location (useful when running unit tests). Instead, a custom
target directory can be set using an absolute path, for instance,
"~/path_to/my_file.ext". In both cases the path specified in \code{data_dir}
does not play a role. Alternatively, if a relative path is provided,
function downloads the data into the the default folder "find.dxgap.data".
Please, make sure the environment variable "DXGAP_DATADIR" is set in your
.Renviron.}
}
\value{
\code{download_hbc()} returns invisibly the file path in which data are
stored.
}
\description{
WHO each five years lists the 20 + 10 countries classified as "High-Burden".
\itemize{
\item The first 20 account for roughly 80\%
of the share of global incidence for a given 5 years period.
\item The last 10 accounts for roughly 2\%-3\% of the share of global incidence for
a given 5 years period.
}
}
\examples{
\dontrun{
download_hbc("who_hbc_2023-07-28.pdf")
}
}
