% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbc.R
\name{hbc}
\alias{hbc}
\alias{download_hbc}
\title{TB High-Burden Countries data from WHO}
\usage{
download_hbc(file_name)
}
\arguments{
\item{file_name}{A string containing the name of the file to be read.}
}
\value{
\code{download_hbc()} returns invisibly the file path in which data are
stored.
}
\description{
WHO each five years lists the 20 + 10 countries classified as "High-Burden".
\itemize{
\item The first 20 account for roughly 80\%
of the share of global incidence for a given 5 years period.
\item The last 10 accounts for roughly 2\%-3\% of the share of global incidence for
a given 5 years period.
}
}
\examples{
\dontrun{
download_hbc("who_2023-07-28_hbc.pdf")
}
}
