% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_dx}
\alias{load_dx}
\title{Load diagnostic gap data}
\usage{
load_dx()
}
\value{
A list of tibbles.
}
\description{
\code{load_dx()} loads all files frames created by \code{\link[=write_data_dir]{write_data_dir()}} into a
list of tibbles.
}
\examples{
\dontrun{
load_dx()
}
}
\seealso{
\code{\link[=build_dm]{build_dm()}}
}
