% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_dx}
\alias{load_dx}
\title{Load diagnostic gap data}
\usage{
load_dx(data_files = list_dx())
}
\arguments{
\item{data_files}{}
}
\description{
\code{load_dx()} loads a list of data frames created by
\code{\link[=write_data_dir]{write_data_dir()}}
}
