% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_tb_dashboard}
\alias{read_tb_dashboard}
\title{Read the TB dashboard}
\usage{
read_tb_dashboard(file_name, data_dir = Sys.getenv("DXGAP_DATADIR"))
}
\arguments{
\item{file_name}{Either an absolute or relative path to the raw Excel TB
dashboard. If a relative path is provided, data is read from the
findtb.data. Instead, if an absolute path is provided, data is read from
any other location.}

\item{data_dir}{An absolute path pointing to the findtb.data folder. By
default, the path is returned by the \code{DXGAP_DATADIR} environment
variable.}
}
\value{
A tibble
}
\description{
\code{read_tb_dashboard()} reads the raw Excel TB dashboard into R, keeping all
the information in an untidied format.
}
\examples{
\dontrun{
read_tb_dashboard()
}
}
\seealso{
\code{\link[=tidy_tb_dashboard]{tidy_tb_dashboard()}} to tidy this dashboard.
}
