% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_report.R
\name{render_bulk}
\alias{render_bulk}
\title{Render templates across multiple years}
\usage{
render_bulk(
  template_name,
  disease,
  estimated = NULL,
  notified = NULL,
  year = NULL,
  vars = NULL,
  override_vars_check = FALSE
)
}
\arguments{
\item{template_name}{String containing the name of the template to render.
Run \code{\link[=view_templates]{view_templates()}} to see a list of valid options.}

\item{disease}{A character of length one identifying the disease for which
the user wants to build a wide table. The tibble \code{dxgap_diseases} shows the
diseases that are currently supported.}

\item{estimated}{Override the default \code{NULL} with a dot-separated character
indicating which field from which table should be used as estimated cases
for DX Gap computation. If kept to \code{NULL}, the function will use the value
\code{estimates_table.estimates_field} from the \code{notified} column in the
\code{dxgap_diseases} meta table.}

\item{notified}{Override the default \code{NULL} with a dot-separated character
indicating which field from which table should be used as notified cases
for DX Gap computation. If kept to \code{NULL}, the function will use the value
\code{notifications_table.notifications_field} from the \code{notified} column in the
\code{dxgap_diseases} meta table.}

\item{year}{An integer vector indicating consecutive years to filter the
data on. Defaults to NULL, returning all years present in the data.}

\item{vars}{A vector of strings naming columns to subset the data on. Passed
to \code{\link[=build_tbl]{build_tbl()}}. Defaults to NULL, indicating all variables should be
used.}

\item{override_vars_check}{Logical indicating whether to override checks on
supported vars. Defaults to FALSE. If TRUE, consistent results are not
guaranteed.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{render_bulk()} is a convenience wrapper around \code{render_report()} that
allows you to render templates across multiple years.
}
\section{Vars Check}{

This check consists in comparing the character vector of variable names
supplied by the user through the \code{vars} argument, with those that were
selected in light of the exploratory data analysis. If some of the provided
variables are not part of the original subset, the function will throw an
error. However, new variables names can always be added overriding the check
by setting \code{override_vars_check = TRUE}. If a new variable name should be
part of the core subset, it should be added to the \code{dxgap_diseases} tibble.
}

\examples{
\dontrun{
tb_vars <- c("year", "is_hbc", "country_code", "dx_gap",
             "pop_total", "pop_urban_perc", "pop_density", "gdp", "c_newinc",
             "e_inc_num", "e_mort_100k", "culture", "smear", "xpert", "m_wrd")
render_bulk("eda.Rmd", disease = "tb", year = 2018:2021, vars = tb_vars)
}
}
