% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_report.R
\name{render_report}
\alias{render_report}
\title{Render a template report}
\usage{
render_report(
  template_name,
  disease,
  estimated = NULL,
  notified = NULL,
  year = NULL,
  vars = NULL,
  interactive = TRUE,
  override_vars_check = FALSE
)
}
\arguments{
\item{disease}{A character of length one identifying the disease for which
the user wants to build a wide table. The tibble \code{dxgap_diseases} shows the
diseases that are currently supported.}

\item{estimated}{Override the default \code{NULL} with a dot-separated character
indicating which field from which table should be used as estimated cases
for DX Gap computation. If kept to \code{NULL}, the function will use the value
\code{estimates_table.estimates_field} from the \code{notified} column in the
\code{dxgap_diseases} meta table.}

\item{notified}{Override the default \code{NULL} with a dot-separated character
indicating which field from which table should be used as notified cases
for DX Gap computation. If kept to \code{NULL}, the function will use the value
\code{notifications_table.notifications_field} from the \code{notified} column in the
\code{dxgap_diseases} meta table.}

\item{year}{Integer matching the year(s) of the report to be rendered. Can be
a single integer like \code{2019}, or a vector of integers such as \code{2019:2021}.
The argument is to passed to \code{build_dm()} under the hood.}

\item{vars}{A vector of strings naming columns to subset the data on.
Defaults to NULL, indicating all variables should be used.}

\item{interactive}{Logical indicating whether to open the report with the
RStudio Viewer.}
}
\description{
\code{render_report()} generates a report for a given template and year. The final
output can be viewed interactively in RStudio. Reports can also be rendered
across multiple years.
}
\examples{
\dontrun{
tb_vars <- c("year", "is_hbc", "country_code", "dx_gap",
             "pop_total", "pop_urban_perc", "pop_density", "gdp", "c_newinc",
             "e_inc_num", "e_mort_100k", "culture", "smear", "xpert", "m_wrd")
render_report(
  template_name = "eda.Rmd",
  disease = "tb",
  estimated = "who_estimates.e_inc_num",
  notified = "who_notifications.c_newinc",
  year = 2019,
  vars = tb_vars
)
}
}
