% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\name{run_mod_lm}
\alias{run_mod_lm}
\title{Run a suite of statistical models}
\usage{
run_mod_lm(
  tbl,
  mod,
  preproc,
  folds,
  metrics,
  rank_metric,
  cross = TRUE,
  seed = 2023
)
}
\arguments{
\item{tbl}{Input data frame containing the data to model.}

\item{mod}{A list of models to run.}

\item{preproc}{A list of pre-processing steps.}

\item{folds}{An integer. The number of cross-validation folds.}

\item{metrics}{A tibble containing the performance metrics to evaluate.}

\item{rank_metric}{A metric from \code{metrics} to rank results by.}

\item{cross}{A logical: should all combinations of the pre-processors and
models be used to create the workflows? If FALSE, the length of preproc and
models should be equal.}

\item{seed}{A single integer.}
}
\value{
A list containing a workset, ranked model results, and a final model
fit.
}
\description{
\code{run_mod_lm()} runs a suite of statistical models, returning a final model fit.
}
\examples{
\dontrun{
tbl <-
  build_tbl("tb", 2019, vars = dxgap_const$tb_vars) |>
  compute_dx_gap() |>
  dplyr::mutate(is_hbc = forcats::as_factor(is_hbc)) |>
  dplyr::select(-any_of(c("year", "country")))

preproc_list <- get_mod_preproc(
  .tbl = tbl,
  .neighbors = 5,
  .threshold = 0.25,
  .impute_vars = c("gdp", "e_inc_num", "pop_total")
 )

mod_list <- get_mod_mod("regression", "lm")

run_mod_lm(
  tbl,
  preproc = preproc_list,
  mod = mod_list,
  folds = 10,
  metrics = yardstick::metric_set(yardstick::rmse, yardstick::rsq),
  rank_metric = "rmse"
)
}
}
