% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb.R
\name{wb}
\alias{wb}
\alias{download_wb}
\alias{read_wb}
\alias{tidy_wb}
\title{Data from World Bank}
\usage{
download_wb(
  file_name = tempfile(compose_file_name("wb", download_date, indicator), fileext =
    ".csv"),
  indicator,
  range_years,
  download_date = as.character(Sys.Date()),
  data_dir = Sys.getenv("DXGAP_DATADIR")
)

read_wb(file_name, data_dir = Sys.getenv("DXGAP_DATADIR"))

tidy_wb(data, year = NULL)
}
\arguments{
\item{file_name}{A string containing the name of the file to be read.}

\item{indicator}{A string indicating the label of the data set as documented
in the World Bank API. For instance, "SP.POP.TOTL".}

\item{range_years}{The range of the years to be downloaded as a string.}

\item{download_date}{A string in the format "\%Y-\%m-\%d". Defaults to today's
date.}

\item{data_dir}{Path containing the directory to read the data from. Defaults
to the path set by the environment variable \code{"DXGAP_DATADIR"}.}

\item{data}{A tibble returned from the corresponding \code{read_()} function.}

\item{year}{A year to filter the data by. Defaults to \code{NULL}, returning data
for all years.}
}
\value{
\code{download_wb()} returns invisibly the file path in which data are
stored.

\code{read_wb()} a tibble containing the data set.

\code{tidy_wb()} a tibble. This is a tidied version of the input tibble.
}
\description{
These help pages document the lower-level API to individually download, read,
and tidy data. For a higher-level API that works across all data sets,
see:
\itemize{
\item \code{\link[=write_data_dir]{write_data_dir()}} to download all data
\item \code{\link[=load_dx]{load_dx()}} to load all data
}
}
\details{
The data sets currently available from WHO in this package are:
\itemize{
\item urban population
\item total population
\item population density
\item gdp
}
}
\section{Source}{

The World Bank API is documented at \url{https://datahelpdesk.worldbank.org/knowledgebase/articles/898581}.
}

\examples{
\dontrun{
pop_urban <- download_wb(
  file_name = compose_date_dataset_file_name("wb", dataset = "pop_urban", file_ext = ".csv"),
  indicator = "SP.URB.TOTL.IN.ZS",
  range_years = "2015:2023"
)
pop_density <-
  download_wb(
    file_name = compose_date_dataset_file_name("wb", dataset = "pop_density", file_ext = ".csv"),
    indicator = "EN.POP.DNST",
    range_years = "2015:2023"
  )
pop_total <- download_wb(
  file_name = compose_date_dataset_file_name("wb", dataset = "pop_total", file_ext = ".csv"),
  indicator = "SP.POP.TOTL",
  range_years = "2015:2023"
)
gdp <- download_wb(
  file_name = compose_date_dataset_file_name("wb", dataset = "gdp", file_ext = ".csv"),
  indicator = "NY.GDP.MKTP.CD",
  range_years = "2015:2023"
)
}
\dontrun{
read_wb("wb_2023-07-28_pop_urban.csv")
}
\dontrun{
read_wb("wb_2023-07-28_pop_urban.csv") |>
  tidy_wb()
}
}
