% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who.R
\name{who}
\alias{who}
\alias{download_who}
\alias{read_who}
\alias{tidy_who}
\title{Data from WHO}
\usage{
download_who(file_name, url_endpoint)

read_who(file_name, data_dir = Sys.getenv("DXGAP_DATADIR"))

tidy_who(data, year = NULL, .shape = "long")
}
\arguments{
\item{file_name}{A string containing the name of the file to be read.}

\item{url_endpoint}{A string indicating the name of the data set to be
downloaded. One of the following:
\itemize{
\item budget
\item community
\item estimates
\item expenditure_utilisation
\item labs
\item notifications
}}

\item{data_dir}{Path containing the directory to read the data from. Defaults
to the path set by the environment variable \code{"DXGAP_DATADIR"}.}

\item{data}{A tibble returned from the corresponding \code{read_()} function.}

\item{year}{A year to filter the data by. Defaults to \code{NULL}, returning data
for all years.}

\item{.shape}{A string indicating if the data should be in "wide", or "long"
format. Defaults to "long".}
}
\value{
\code{download_who()} returns invisibly the file path in which data are
stored.

\code{read_who()} returns a tibble containing the data set.

\code{tidy_who()} returns s tibble. This is a tidied version of the input
tibble.
}
\description{
These help pages document the lower-level API to individually download, read,
and tidy data. For a higher-level API that works across all data sets,
see:
\itemize{
\item \code{\link[=write_data_dir]{write_data_dir()}} to download all data
}
}
\details{
The data sets currently available from WHO in this package are:
\itemize{
\item notifications
\item estimates
\item budget
\item community
\item expenditures
\item laboratories
\item hbc
}
}
\examples{
\dontrun{
notification <- download_who(
  file_name = compose_date_dataset_file_name("who", dataset = "notifications", file_ext = ".csv"),
  url_endpoint = "notifications"
)
estimates <- download_who(
  file_name = compose_date_dataset_file_name("who", dataset = "estimates", file_ext = ".csv"),
  url_endpoint = "estimates"
)
budget <- download_who(
  file_name = compose_date_dataset_file_name("who", dataset = "budget", file_ext = ".csv"),
  url_endpoint = "budget"
)
community_engagement <- download_who(
  file_name = compose_date_dataset_file_name("who", dataset = "community", file_ext = ".csv"),
  url_endpoint = "community"
)
expenditure_and_utilisation <- download_who(
  file_name = compose_date_dataset_file_name("who", dataset = "expenditures", file_ext = ".csv"),
  url_endpoint = "expenditure_utilisation"
)
laboratories <- download_who(
  file_name = compose_date_dataset_file_name("who", dataset = "laboratories", file_ext = ".csv"),
  url_endpoint = "labs"
)
}
\dontrun{
read_who("who_2023-08-30_laboratories.csv")
read_who("who_2023-07-28_hbc.csv")
}
\dontrun{
read_who("who_2023-08-30_laboratories.csv") |>
  tidy_who()
}
}
