% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{write_data_dir}
\alias{write_data_dir}
\title{Create folder containing data for the project.}
\usage{
write_data_dir(
  path = fs::path_dir(getwd()),
  dirs = c("who", "global-fund", "world-bank", "meta", "report")
)
}
\arguments{
\item{path}{A path where the folder should be created. By default, it is
created at the same level as the find.dxgap root directory.}

\item{dirs}{a character vector with sub-folder names.}
}
\value{
the path to the folder, invisibly.
}
\description{
This function has side effects. It creates a folder storing data for the
project. In addition it sets the \code{DXGAP_DATADIR} environment variable which
is needed for the \verb{read_*()} family of functions to work, in particular when
running tests.
}
\examples{
write_data_dir() # Create folder at same level as project
write_data_dir("~/Documents/my_projects") # Specify specific folder
}
